third_party/blink/web_tests/platform/generic/fast/forms/fieldset/fieldset-form-collection-radionode-list-expected.txt

This test ensures fieldset.elements does NOT contain RadioNodeList.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS owner.elements.length is 23
PASS notRadioNodeList instanceof RadioNodeList is false
PASS successfullyParsed is true

TEST COMPLETE
